Rename module namespace `Cardano.Wallet.Write`.
This commit renames the above module namespace to `Cardano.Write`.
