+++ lib/balance-tx/cardano-balance-tx.cabal
@@ -80,16 +80,16 @@ library
     Cardano.Tx.Balance.Internal.CoinSelection
     Cardano.Tx.Balance.Internal.CoinSelection.Gen
     Cardano.Wallet.Shelley.Compatibility.Ledger
-    Cardano.Wallet.Write.ProtocolParameters
-    Cardano.Wallet.Write.Tx
-    Cardano.Wallet.Write.Tx.Balance
-    Cardano.Wallet.Write.Tx.Balance.TokenBundleSize
-    Cardano.Wallet.Write.Tx.Gen
-    Cardano.Wallet.Write.Tx.Redeemers
-    Cardano.Wallet.Write.Tx.Sign
-    Cardano.Wallet.Write.Tx.SizeEstimation
-    Cardano.Wallet.Write.Tx.TimeTranslation
-    Cardano.Wallet.Write.UTxOAssumptions
+    Cardano.Write.ProtocolParameters
+    Cardano.Write.Tx
+    Cardano.Write.Tx.Balance
+    Cardano.Write.Tx.Balance.TokenBundleSize
+    Cardano.Write.Tx.Gen
+    Cardano.Write.Tx.Redeemers
+    Cardano.Write.Tx.Sign
+    Cardano.Write.Tx.SizeEstimation
+    Cardano.Write.Tx.TimeTranslation
+    Cardano.Write.UTxOAssumptions
 
 test-suite test
   import:             language, opts-exe
