+++ lib/balance-tx/lib/Cardano/Write/Tx/Balance.hs
@@ -23,7 +23,7 @@
 {-# OPTIONS_GHC -fno-warn-ambiguous-fields #-}
 #endif
 
-module Cardano.Wallet.Write.Tx.Balance
+module Cardano.Write.Tx.Balance
     (
     -- * Balancing transactions
       balanceTransaction
@@ -117,9 +117,9 @@ import Cardano.Wallet.Primitive.Types.Tx
     ( SealedTx, sealedTxFromCardano )
 import Cardano.Wallet.Primitive.Types.Tx.Constraints
     ( TxSize (..), txOutMaxCoin, txOutMaxTokenQuantity )
-import Cardano.Wallet.Write.ProtocolParameters
+import Cardano.Write.ProtocolParameters
     ( ProtocolParameters (..) )
-import Cardano.Wallet.Write.Tx
+import Cardano.Write.Tx
     ( Address
     , Coin (..)
     , FeePerByte (..)
@@ -149,17 +149,17 @@ import Cardano.Wallet.Write.Tx
     , txBody
     , withConstraints
     )
-import Cardano.Wallet.Write.Tx.Balance.TokenBundleSize
+import Cardano.Write.Tx.Balance.TokenBundleSize
     ( mkTokenBundleSizeAssessor )
-import Cardano.Wallet.Write.Tx.Redeemers
+import Cardano.Write.Tx.Redeemers
     ( ErrAssignRedeemers (..), Redeemer (..), assignScriptRedeemers )
-import Cardano.Wallet.Write.Tx.Sign
+import Cardano.Write.Tx.Sign
     ( estimateKeyWitnessCount, estimateSignedTxSize )
-import Cardano.Wallet.Write.Tx.SizeEstimation
+import Cardano.Write.Tx.SizeEstimation
     ( TxSkeleton (..), assumedTxWitnessTag, estimateTxCost )
-import Cardano.Wallet.Write.Tx.TimeTranslation
+import Cardano.Write.Tx.TimeTranslation
     ( TimeTranslation )
-import Cardano.Wallet.Write.UTxOAssumptions
+import Cardano.Write.UTxOAssumptions
     ( UTxOAssumptions (..), assumedInputScriptTemplate )
 import Control.Arrow
     ( left )
