+++ lib/wallet/cardano-wallet.cabal
@@ -922,8 +922,8 @@ test-suite unit
     Cardano.Wallet.Submissions.OperationsSpec
     Cardano.Wallet.Submissions.PrimitivesSpec
     Cardano.Wallet.TokenMetadataSpec
-    Cardano.Wallet.Write.Tx.Balance.TokenBundleSizeSpec
-    Cardano.Wallet.Write.TxSpec
+    Cardano.Write.Tx.Balance.TokenBundleSizeSpec
+    Cardano.Write.TxSpec
     Cardano.WalletSpec
     Control.Concurrent.ConciergeSpec
     Control.Monad.UtilSpec
