+++ lib/wallet/src/Cardano/Wallet/Address/Discovery/RandomAny.hs
@@ -81,7 +81,7 @@ import System.Random
     ( mkStdGen )
 
 import qualified Cardano.Wallet.Address.Discovery.Random as Rnd
-import qualified Cardano.Wallet.Write.UTxOAssumptions as UTxOAssumptions
+import qualified Cardano.Write.UTxOAssumptions as UTxOAssumptions
 
 --------------------------------------------------------------------------------
 --
