+++ lib/wallet/src/Cardano/Wallet/Byron/Compatibility.hs
@@ -103,7 +103,7 @@ import qualified Cardano.Wallet.Primitive.Types.Tx as W
 import qualified Cardano.Wallet.Primitive.Types.Tx.Constraints as W
 import qualified Cardano.Wallet.Primitive.Types.Tx.TxOut as W
     ( TxOut (TxOut) )
-import qualified Cardano.Wallet.Write.Tx as Write
+import qualified Cardano.Write.Tx as Write
 import qualified Data.Map.Strict as Map
 import qualified Ouroboros.Consensus.Block as O
